+++ src/main/java/de/expeehaa/spigot/config/Wand.java
@@ -8,7 +8,7 @@ import org.bukkit.potion.PotionEffect;
 public class Wand {
 	
 	public Material material;
-	public short damaged;
+	public int damaged;
 	public boolean rightClick;
 	
 	public List<PotionEffect> potionEffectList;
@@ -17,7 +17,7 @@ public class Wand {
 	
 	public double damage;
 	
-	public Wand(Material material, short damaged, List<PotionEffect> potionEffectList, double speed, boolean rightClick, double damage) {
+	public Wand(Material material, int damaged, List<PotionEffect> potionEffectList, double speed, boolean rightClick, double damage) {
 		this.material = material;
 		this.damaged = damaged;
 		this.potionEffectList = potionEffectList;
